% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_entry.R
\name{find_entry}
\alias{find_entry}
\title{Find an ontology term based on value of an element}
\usage{
find_entry(ontology, element, string)
}
\arguments{
\item{ontology}{a list (of class ontology) in which to look}

\item{element}{in which element of the ontology should the search be done, in
disease ontology, you could e.g. provide "xref" to look at cross-references, or
"id" to look at specific entries}

\item{string}{a character string to search for, e.g. "ICD10CM:I10".}
}
\description{
Searches for a string in an element of an ontology and returns
a logical vector, which could be used to subset the ontology.
}
\author{
Daniel Lindholm
}

