% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_depth.R
\name{get_depth}
\alias{get_depth}
\title{Get position in tree for a specific term}
\usage{
get_depth(ontology, id, root = "DOID:4")
}
\arguments{
\item{ontology}{A list (of class ontology) in which to look}

\item{id}{The identifier of the term to look up (character string)}

\item{root}{Identifier of the root.}
}
\value{
Outputs a numeric value, corresponding to the depth, where 1 is the root.
}
\description{
Looks up where in the ontology tree (at which depth) a term is.
}
\author{
Daniel Lindholm
}

