% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parents.R
\name{get_parents}
\alias{get_parents}
\title{Get parent nodes of a specific ontology term}
\usage{
get_parents(ontology, id, root = "DOID:4", include_entry = FALSE)
}
\arguments{
\item{ontology}{A list (of class ontology) in which to look}

\item{id}{The identifier of the term to look up (character string)}

\item{root}{The identifier of the root entry. Defaults to "DOID:4", which is
the term "disease" in the disease ontology (http://www.disease-ontology.org/).}

\item{include_entry}{Should the term to look up be included in the results?
Defaults to FALSE}
}
\value{
Outputs a dataframe with identifiers, descriptions, and depth.
}
\description{
Will look up in the hierarchy until the root is reached.
}
\author{
Daniel Lindholm
}

