% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_id.R
\name{map_to_id}
\alias{map_to_id}
\title{Map ICD, OMIM, SNOMED or any cross-reference code to an ontology identifier}
\usage{
map_to_id(ontology, code, from = "ICD10CM", root = "DOID:4")
}
\arguments{
\item{ontology}{a list (of class ontology) in which to look}

\item{code}{the code to look-up}

\item{from}{Which classification is 'code' apart of? Defaults to "ICD10CM",
other options (when using the disease ontology database) include "OMIM", "MSH",
"SNOMEDCT_US_2016-03-01", etc.}
}
\value{
Outputs a character string containing the mapped identifier.
}
\description{
Will look up a code within the "xref" part of an ontology database
and translate to an ontology identifier. If several identifiers are found, the
one that is deepest in the tree will be returned. If there are multiple entries
within the same depth, the first identifier in the list will be returned.
}
\author{
Daniel Lindholm
}

