% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_obo.R
\name{read_obo}
\alias{read_obo}
\title{Read and parse obo (Open Biomedical Ontologies) files}
\usage{
read_obo(obo_file)
}
\arguments{
\item{obo_file}{a character vector pointing to the obo file to read.}
}
\value{
Outputs a list of class 'ontology'.
}
\description{
Reads and parses obo files.
}
\examples{
# This example downloads the most recent disease ontology obo file
# and converts it to an R list:

doid <- read_obo("http://purl.obolibrary.org/obo/doid.obo")
head(doid)

}
\author{
Daniel Lindholm
}

