% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_obsolete.R
\name{remove_obsolete}
\alias{remove_obsolete}
\title{Remove obsolete terms from an ontology}
\usage{
remove_obsolete(ontology)
}
\arguments{
\item{ontology}{a list (of class ontology) in which to look}
}
\value{
Outputs a list with obsolete terms removed.
}
\description{
Convenience function for removing obsolete terms from an ontology.
}
\author{
Daniel Lindholm
}

